Moved CasemNodeComparator into separate file
